// src/usecase/customer/update/update.customer.dto.ts
export interface IInputUpdateCustomerDTO {
    id: string;
    name: string;
    address: {
        street: string;
        number: number;
        zip: string;
        city: string;
    }
}

export interface IOutputUpdateCustomerDTO {
    id: string;
    name: string;
    address: {
        street: string;
        number: number;
        zip: string;
        city: string;
    }
}